requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
requests==2.32.3
python-dotenv==1.0.1
apscheduler==3.10.4
